Theremin enhancements to sonify all 3 axes (#78)
* added UI and functionality sonify different axes and modify sonification properties.

* starting development

* initial commit

* (broken) version of the theramin for a light sensor

* "working" with other types of sensors but not light

This has a partial interface constructed -- I got as far as showing a
list of services. When I use
    const lightSensors = useServices({ serviceClass: SRV_REFLECTED_LIGHT})

lightSensors is empty. If I use something else, like SRV_ACCELEROMETER
or SRV_BUTTON, the result of useServices is not empty. Until I debug
that I cannot ensure that a light sensor is selected

* added james suggestion for the right imports.

* Working light "theramin"

* formatting fixes and modulo fix.

Co-authored-by: Jennifer Mankoff <jmankoff@cs.washington.edu>